Refactored preprocessing
- Added new cleanup logic
- Switched from using re to using regex for advanced character removal
- Changed function structure to use pandas dataframe applymap for better performance
